% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fiscrape}
\alias{fiscrape}
\title{Run the screen scraper}
\usage{
fiscrape(update_bdays = FALSE, debug = FALSE)
}
\arguments{
\item{update_bdays}{boolean}

\item{debug}{boolean}
}
\description{
The main function
}
