% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_race.R
\name{update_person}
\alias{update_person}
\title{Update Person Info}
\usage{
update_person(target_gender, target_compid, target_fisid, ...)
}
\arguments{
\item{target_gender}{character}

\item{target_compid}{integer}

\item{target_fisid}{character}

\item{...}{named changes, e.g. \code{length = 30} or \code{cat1 = "WC"}}
}
\description{
Update race info in both remote and local
databases for a specified raceid.
}
