% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_race.R
\name{update_race}
\alias{update_race}
\title{Update Race Info}
\usage{
update_race(raceid, ...)
}
\arguments{
\item{raceid}{integer}

\item{...}{named changes, e.g. \code{length = 30} or \code{cat1 = "WC"}}
}
\description{
Update race info in both remote and local
databases for a specified raceid.
}

