% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_xc_fac.R
\name{update_xc_fac}
\alias{update_xc_fac}
\title{Create Updated Version of XC_FAC}
\usage{
update_xc_fac()
}
\value{
A data frame
}
\description{
Creates an updated version of xc_fac used for
standardizing median percent back by gender and
race type.
}

