% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{db_xc_local}
\alias{reexports}
\alias{db_xc_remote}
\alias{reexports}
\alias{ss_query}
\title{Create Local XC DB Connection}
\description{
Create Local XC DB Connection

Create Remote XC DB Connection
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{statskier2}{\code{\link[statskier2]{db_xc_local}}, \code{\link[statskier2]{db_xc_remote}}, \code{\link[statskier2]{ss_query}}}
}}

